Move NodeDetails to feed directory
